requirements.txt
redis
requests
python-dotenv
sendgrid
fastapi
uvicorn
websockets
